Erro PROPOSTA_ID_PROPOSTA

+++ src/main/java/com/ecoswap/ecoswap/services/PropostaRepository.java
@@ -3,5 +3,5 @@ package com.ecoswap.ecoswap.services;
 import com.ecoswap.ecoswap.domain.Proposta;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PropostaRepository extends JpaRepository<Proposta, Integer> {
+public interface PropostaRepository extends JpaRepository<Proposta, Long> {
 }
