+++ src/main/java/com/ecoswap/ecoswap/services/TrocaRepository.java
@@ -3,5 +3,5 @@ package com.ecoswap.ecoswap.services;
 import com.ecoswap.ecoswap.domain.Troca;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TrocaRepository extends JpaRepository<Troca, Integer> {
+public interface TrocaRepository extends JpaRepository<Troca, Long> {
 }
